+++ CHANGELOG.md
@@ -101,6 +101,7 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 - Resolving of Math Subject Classification (MSC) needs to be activated in the references now. [#15883](https://github.com/JabRef/jabref/pull/15883)
 - We improved JabRef's startup performance by loading MSC codes only when activated. [#15883](https://github.com/JabRef/jabref/pull/15883)
 - Embedded postgres is no longer started with JabRef unless "Experimental search (Postgres)" is enabled in General preferences. [#12844](https://github.com/JabRef/jabref/issues/12844)
+- We renamed the sort dropdown option in "Search for unlinked local files" from "Default" to "Grouped by directory". [#16158](https://github.com/JabRef/jabref/issues/16158)
 
 ### Fixed
 
@@ -192,6 +193,10 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 - We fixed handling of `exit` in the LSP server. [#16268](https://github.com/JabRef/jabref/pull/16268)
 - We fixed an issue where `LinkedFile.isOnlineLink()` did not recognize `ftp://` links as online links. [#16400](https://github.com/JabRef/jabref/issues/16400)
 - We fixed an issue where deleting an entry removed the group filter and defaulted to the home view. [#16036](https://github.com/JabRef/jabref/issues/16036)
+- We fixed the "Search for unlinked local files" that was not applying JabRef's theme. [#16158](https://github.com/JabRef/jabref/issues/16158)
+- We fixed the question-mark icon appearing in the "Search for unlinked local files" wizard pages. [#16158](https://github.com/JabRef/jabref/issues/16158)
+- We fixed the "select files to import" header, it no longer shows during the file scan. [#16158](https://github.com/JabRef/jabref/issues/16158)
+- We fixed the space key not toggling the file checkboxes after we scan the document in "Search for unlinked local files". [#16158](https://github.com/JabRef/jabref/issues/16158)
 
 ### Removed